backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
distro==1.9.0
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-aiplatform==1.141.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-cloud-texttospeech==2.34.0
google-crc32c==1.8.0
google-genai==1.66.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
